Project 012 - Audio Summarizer/requirements.txt
streamlit
python-dotenv
google-generativeai
SpeechRecognition
pydub
Ipython
